+++ test/testUcumLhcUtils.spec.js
@@ -94,6 +94,15 @@ describe('Test validateUnitString method', function() {
     assert.equal(utils.validateUnitString('-123').status, 'invalid');
   });
 
+  it('should say a unit that is only an annotation is valid - {g}', ()=>{
+    // {g} was getting marked invalid, along with the suggestion that maybe the
+    // user meant [g].  That is a useful suggestion, but it should be valid.
+    const result = utils.validateUnitString('{g}');
+    assert.equal(result.status, 'valid');
+    assert(result.msg.length > 0);
+  });
+
+
 }); // end validateUnitString tests
 
 
@@ -297,6 +306,36 @@ it("should return a message for invalid unit strings", function() {
     assert.equal(resp.toVal.toPrecision(7), -16.66667);
   });
 
+  it('should not convert {degF} to Cel', ()=>{
+    const resp = utils.convertUnitTo('{degF}', 5, 'Cel');
+    assert.equal(resp.status, 'failed');
+    assert.equal(
+      '{degF} is a valid unit expression, but did you mean [degF] (degrees Fahrenheit)?',
+      resp.msg[0]);
+    assert.equal('Sorry.  {degF} cannot be converted to Cel.', resp.msg[1]);
+  });
+
+  it('should convert {degF} to 1', ()=>{
+    const resp = utils.convertUnitTo('{degF}', 2, '1');
+    assert.equal(resp.status, 'succeeded');
+    assert.equal(resp.toVal, 2);
+    assert.equal(resp.msg[0],
+      '{degF} is a valid unit expression, but did you mean [degF] (degrees Fahrenheit)?');
+  });
+
+  it('should convert {aaa} to {bbb}', ()=>{
+    const resp = utils.convertUnitTo('{aaa}', 2, '{bbb}');
+    assert.equal(resp.status, 'succeeded');
+    assert.equal(resp.toVal, 2);
+    assert.equal(resp.msg.length, 0);
+  });
+
+  it('should convert 3.{aaa} to {bbb}', ()=>{
+    const resp = utils.convertUnitTo('3.{aaa}', 2, '{bbb}');
+    assert.equal(resp.status, 'succeeded');
+    assert.equal(resp.toVal, 6);
+    assert.equal(resp.msg.length, 0);
+  });
 }); // end convertUnitTo tests
 
 
@@ -338,18 +377,9 @@ describe('Test getSynonyms method', function() {
 
 
 describe('convertToBaseUnits', ()=> {
-  // Base units currently appear in the dimension vector in this order
-  // (taken from UnitTables.getInstance().dimVecIndexToBaseUnit_)
-  /*   '0': 'm',
-       '1': 's',
-       '2': 'g',
-       '3': 'rad',
-       '4': 'K',
-       '5': 'C',
-       '6': 'cd' */
-
   it('should convert cm2/ms3 to base units', ()=>{
     const baseUnitData = utils.convertToBaseUnits('cm2/ms3', 1);
+    assert.equal(baseUnitData.status, 'succeeded');
     assert.equal(baseUnitData.msg.length, 0);
     assert.equal(baseUnitData.magnitude, 100000);
     assert.equal(baseUnitData.fromUnitIsSpecial, false);
@@ -360,6 +390,7 @@ describe('convertToBaseUnits', ()=> {
   it('should convert B[10.nV] to base units', ()=>{
     // Equivalent to 0.0000316227766016838 g.m2/(C.s2); 1 B[10.nV]=2*log10(10.nV)
     const baseUnitData = utils.convertToBaseUnits('B[10.nV]', 1);
+    assert.equal(baseUnitData.status, 'succeeded');
     assert.equal(baseUnitData.msg.length, 0);
     assert.equal(baseUnitData.magnitude.toPrecision(15), 0.0000316227766016838);
     assert.equal(baseUnitData.fromUnitIsSpecial, true);
@@ -368,6 +399,7 @@ describe('convertToBaseUnits', ()=> {
 
   it('should convert [degF] to base units', ()=>{
     const baseUnitData = utils.convertToBaseUnits('[degF]', 32);
+    assert.equal(baseUnitData.status, 'succeeded');
     assert.equal(baseUnitData.msg.length, 0);
     assert.equal(baseUnitData.magnitude.toPrecision(15), 273.15);
     assert.equal(baseUnitData.fromUnitIsSpecial, true);
@@ -376,6 +408,7 @@ describe('convertToBaseUnits', ()=> {
 
   it('should convert 2.[degF] to base units', ()=>{
     const baseUnitData = utils.convertToBaseUnits('2.[degF]', 16);
+    assert.equal(baseUnitData.status, 'succeeded');
     assert.equal(baseUnitData.msg.length, 0);
     // Special units do not get a magnitude returned
     assert.equal(baseUnitData.magnitude.toPrecision(15), 273.15);
@@ -385,6 +418,7 @@ describe('convertToBaseUnits', ()=> {
 
   it('should convert [degF].2 to base units', ()=>{
     const baseUnitData = utils.convertToBaseUnits('[degF].2', 16);
+    assert.equal(baseUnitData.status, 'succeeded');
     assert.equal(baseUnitData.msg.length, 0);
     assert.equal(baseUnitData.magnitude.toPrecision(15), 273.15);
     assert.equal(baseUnitData.fromUnitIsSpecial, true);
@@ -393,16 +427,26 @@ describe('convertToBaseUnits', ()=> {
 
   it('should convert [degF] to base units with a different value', ()=>{
     const baseUnitData = utils.convertToBaseUnits('[degF]', 41);
+    assert.equal(baseUnitData.status, 'succeeded');
     assert.equal(baseUnitData.msg.length, 0);
     assert.equal(baseUnitData.magnitude.toPrecision(15), 278.15);
     assert.equal(baseUnitData.fromUnitIsSpecial, true);
     assert.deepEqual(baseUnitData.unitToExp, {'K': 1});
   });
 
+  it('should convert {degF} to 1', ()=>{
+    const baseUnitData = utils.convertToBaseUnits('{degF}', 2);
+    assert.equal(baseUnitData.status, 'succeeded');
+    assert.equal(baseUnitData.msg.length, 1);
+    assert.equal(baseUnitData.magnitude, 2);
+    assert.equal(baseUnitData.fromUnitIsSpecial, false);
+  });
+
   it('should not covert [iU] to base units', ()=>{
     // [iU] is an "arbitrary" unit and cannot be converted
     const baseUnitData = utils.convertToBaseUnits('[iU]', 1);
     assert(baseUnitData.msg.length > 0);
+    assert.equal(baseUnitData.status, 'failed');
     assert.equal(baseUnitData.magnitude, undefined);
     assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);
@@ -412,6 +456,7 @@ describe('convertToBaseUnits', ()=> {
     // [iU] is an "arbitrary" unit and cannot be converted
     const baseUnitData = utils.convertToBaseUnits('m.[iU]', 1);
     assert(baseUnitData.msg.length > 0);
+    assert.equal(baseUnitData.status, 'failed');
     assert.equal(baseUnitData.magnitude, undefined);
     assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);
@@ -420,6 +465,7 @@ describe('convertToBaseUnits', ()=> {
   it('should return an error message for a unit string that is a synonym', ()=>{
     const baseUnitData = utils.convertToBaseUnits('foot', 1);
     assert(baseUnitData.msg.length > 0);
+    assert.equal(baseUnitData.status, 'invalid');
     assert.equal(baseUnitData.magnitude, undefined);
     assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);
@@ -428,6 +474,7 @@ describe('convertToBaseUnits', ()=> {
   it('should return an error message for a unit string that is a unrecognized', ()=>{
     const baseUnitData = utils.convertToBaseUnits('zzz', 1);
     assert(baseUnitData.msg.length > 0);
+    assert.equal(baseUnitData.status, 'invalid');
     assert.equal(baseUnitData.magnitude, undefined);
     assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);